Re-write of Estimator logging proceudre
Was previously creating a temp. file within nmrspy/logs/ and
logging method calls to it. This causes errors if an Estimator
instance is loaded using a different nmrespy installation, as the
supporting logfile is not present.

To overcome this, I have changed the procedure so that an attribute with the log text
is simply stored and appended to.